linux-user: support loading scripts with shebang (#!)
The interpreter is assumed to be compatible with the target architecture.

The script loading logic is taken from Linux source code to match logic as closely as possible.

An interpreter can itself be a script (#!/other.script), and thus load another interpreter.
This happens in a loop therefore the loading chain of interpreter-scripts is limited to 5 like in Linux.

Warning: there might be issues with m68k, mips, and mips64 architectures
since the cpu_model returned by those architectures (see linux-user/$arch/target_elf.h)
is dependent on the ELF header of the payload, but in this case the payload
is a script and not a binary.
	This could be fixed either by moving the loading logic or
parts of it to before the cpu_model is set, so that the final ELF binary is available.
An alternative fix is to avoid the loop altogether and call qemu binary again with different arguments.
The downside is that it would require one extra exec syscall per interpreter.

Signed-off-by: Tibor Vass <tibor@docker.com>